Added more relationships

+++ Services/CharacterService/CharacterService.cs
@@ -1,7 +1,9 @@
+using System.Security.Claims;
 using AutoMapper;
 using dotnet_rpg.Data;
 using dotnet_rpg.Dtos.Character;
 using dotnet_rpg.Models;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.EntityFrameworkCore;
 
 namespace dotnet_rpg.Services.CharacterService;
@@ -10,13 +12,19 @@ public class CharacterService : ICharacterService
 {
     private readonly IMapper _mapper;
     private readonly DataContext _context;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public CharacterService(IMapper mapper, DataContext context)
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
         _context = context;
         _mapper = mapper;
+        _httpContextAccessor = httpContextAccessor;
     }
-
+    
+    //get user ID
+    private int GetUserId() =>
+        int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+    
     private static List<Character> characters = new List<Character>
     {
         new Character(),
@@ -32,10 +40,10 @@ public class CharacterService : ICharacterService
         }
     };
 
-    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int id)
+    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
     {
         ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-        var dbCharacters = await _context.Characters.Where(c => c.Id == id).ToListAsync();
+        var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
         serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(dbCharacters);
         return serviceResponse;
 
@@ -54,7 +62,7 @@ public class CharacterService : ICharacterService
 
         try
         {
-            var dbCharacter = await _context.Characters.Where(c => c.Id == id).FirstAsync();
+            var dbCharacter = await _context.Characters.Where(c => c.Id == id && c.User.Id == GetUserId()).FirstAsync();
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
         }
         catch (Exception e)
@@ -71,12 +79,18 @@ public class CharacterService : ICharacterService
         ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
 
         var character = _mapper.Map<Character>(newCharacter);
+        character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
 
         await _context.Characters.AddAsync(character);
         await _context.SaveChangesAsync();
 
-        serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(characters);
-        return serviceResponse;
+        // serviceResponse.Data = _mapper
+        //     .Map<List<GetCharacterDto>>(_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+        //
+        //serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(characters);
+        //serviceResponse.Data = _context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+
+        return await GetAllCharacters();
     }
 
     public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto updatedCharacter)
@@ -86,20 +100,38 @@ public class CharacterService : ICharacterService
         try
         {
             //var character = characters.Where(c => c.Id == updatedCharacter.Id).FirstOrDefault();
-            var character = await _context.Characters.Where(c => c.Id == updatedCharacter.Id).FirstOrDefaultAsync();
-            character.Name = updatedCharacter.Name;
-            character.Class = updatedCharacter.Class;
-            character.Defense = updatedCharacter.Defense;
-            character.HitPoints = updatedCharacter.HitPoints;
-            character.Intelligence = updatedCharacter.Intelligence;
-            character.Strength = updatedCharacter.Strength;
             
-           // _mapper.Map(updatedCharacter, character);
+            //we must include the user since EF do not know anything about it 
+            
+           // var character = await _context.Characters.Include(c => c.User).Where(c => c.Id == updatedCharacter.Id && c.User.Id == GetUserId()).FirstOrDefaultAsync();
 
-            _context.Characters.Update(character);
-            await _context.SaveChangesAsync();
+            Character character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
 
-          
+            
+            if (character.User.Id == GetUserId())
+            {
+
+                character.Name = updatedCharacter.Name;
+                character.Class = updatedCharacter.Class;
+                character.Defense = updatedCharacter.Defense;
+                character.HitPoints = updatedCharacter.HitPoints;
+                character.Intelligence = updatedCharacter.Intelligence;
+                character.Strength = updatedCharacter.Strength;
+
+                // _mapper.Map(updatedCharacter, character);
+
+                _context.Characters.Update(character);
+                await _context.SaveChangesAsync();
+
+            }
+
+            else
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = "Character has not been found!";
+            }
+
+
             //better practice since we don't have to map every single property!
 
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
@@ -120,12 +152,20 @@ public class CharacterService : ICharacterService
         try
         {
             //var character = characters.Where(c => c.Id == id).First();
-            var character = await _context.Characters.Where(c => c.Id == id).FirstAsync();
+            var character = await _context.Characters.Where(c => c.Id == id && c.User.Id == GetUserId()).FirstOrDefaultAsync();
 
-            _context.Characters.Remove(character);
-            await _context.SaveChangesAsync();
+            if (character != null)
+            {
+                _context.Characters.Remove(character);
+                await _context.SaveChangesAsync();
 
-            serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(_context.Characters.ToList());
+                serviceResponse.Data = _mapper.Map<List<GetCharacterDto>>(_context.Characters.ToList());
+            }
+            else
+            {
+                serviceResponse.Success = false;
+                serviceResponse.Message = "Character has not been found!";
+            }
         }
         catch (Exception e)
         {
